Fix a db typing thingy, update spectrumapp tests slightly so that they work again
